pyproject.toml
[project]
name = "napcrawler"
version = "1.0.0"
description = ""
authors = [
    {name = "sakagami0615",email = "sakagami0615@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pytest-playwright (>=0.7.0,<0.8.0)"
]

[tool.poetry]
packages = [{include = "napcrawler", from = "."}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.3"

[project.scripts]
napcrawler = "napcrawler.app:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
indent-width = 4

# フォーマッター設定（blackと同じ整形ルールを適用）
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "I",   # isort (import order)
    "UP",  # pyupgrade
    "S",   # flake8-bandit (セキュリティ)
]
ignore = [
    "E501",  # 行の長さ制限を無視（flake8と合わせるなら無効）
]

# import順の自動整形用（isort相当）
[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false
